So ...I made an eventful call for the map
I haven't figured out how to call more events as the regular page
loads yet.  So for the sake of just having something to show more
of I made a call that calls for a max size of 25, then plots on the
map.

I figure that this way it looks less sparce.
The regular call still works and only loads in 10 at a time so the
inifinite scrolling is still super clutch.

eh, so many attempts at trying to get the other way working...
I just did this as a light patch... for the time being...
just to have a little more pins plotted